software/requirements.txt
# =============================================================================
# `requirements.txt` installs all dependant packages
# =============================================================================
virtualenv
setuptools
wheel
asyncio
asyncua
cryptography
wifi
dataclasses
#rpi-gpio
adafruit-circuitpython-busdevice
adafruit-circuitpython-am2320
adafruit-circuitpython-dht
azure.iot.device
azure.identity
azure.keyvault.certificates
azure.keyvault.secrets
azure.keyvault.keys